+++ uuid/uuid.go
@@ -26,26 +26,26 @@ var (
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-// GenUUID generates v4 UUID (Universally Unique Identifier)
-func GenUUID() string {
-	return GenUUID4()
-}
+// UUID contains UUID data
+type UUID []byte
 
-// GenUUID4 generates random generated UUID
-func GenUUID4() string {
-	uuid := make([]byte, 16)
+// ////////////////////////////////////////////////////////////////////////////////// //
+
+// UUID4 generates random generated UUID v4
+func UUID4() UUID {
+	uuid := make(UUID, 16)
 
 	rand.Read(uuid)
 
-	uuid[6] = (uuid[6] & 0X0F) | 0X40
-	uuid[8] = (uuid[8] & 0X3F) | 0X80
+	uuid[6] = (uuid[6] & 0x0F) | 0x40
+	uuid[8] = (uuid[8] & 0x3F) | 0x80
 
-	return toString(uuid)
+	return UUID(uuid)
 }
 
-// GenUUID5 generates UUID based on SHA-1 hash of namespace UUID and name
-func GenUUID5(ns []byte, name string) string {
-	uuid := make([]byte, 16)
+// UUID5 generates UUID v5 based on SHA-1 hash of namespace UUID and name
+func UUID5(ns []byte, name string) UUID {
+	uuid := make(UUID, 16)
 
 	hash := sha1.New()
 	hash.Write(ns)
@@ -53,26 +53,52 @@ func GenUUID5(ns []byte, name string) string {
 
 	copy(uuid, hash.Sum(nil))
 
-	uuid[6] = (uuid[6] & 0X0F) | 0x50
-	uuid[8] = (uuid[8] & 0X3F) | 0x80
+	uuid[6] = (uuid[6] & 0x0F) | 0x50
+	uuid[8] = (uuid[8] & 0x3F) | 0x80
 
-	return toString(uuid)
+	return UUID(uuid)
 }
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-func toString(uuid []byte) string {
+// String returns string representation of UUID
+func (u UUID) String() string {
 	buf := make([]byte, 36)
 
-	hex.Encode(buf[0:8], uuid[0:4])
+	hex.Encode(buf[0:8], u[0:4])
 	buf[8] = '-'
-	hex.Encode(buf[9:13], uuid[4:6])
+	hex.Encode(buf[9:13], u[4:6])
 	buf[13] = '-'
-	hex.Encode(buf[14:18], uuid[6:8])
+	hex.Encode(buf[14:18], u[6:8])
 	buf[18] = '-'
-	hex.Encode(buf[19:23], uuid[8:10])
+	hex.Encode(buf[19:23], u[8:10])
 	buf[23] = '-'
-	hex.Encode(buf[24:], uuid[10:])
+	hex.Encode(buf[24:], u[10:])
 
 	return string(buf)
 }
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
+// GenUUID generates v4 UUID (Universally Unique Identifier)
+//
+// Deprecated: Use method UUID4.String() instead
+func GenUUID() string {
+	return UUID4().String()
+}
+
+// GenUUID4 generates random generated UUID
+//
+// Deprecated: Use method UUID4.String() instead
+func GenUUID4() string {
+	return UUID4().String()
+}
+
+// GenUUID5 generates UUID based on SHA-1 hash of namespace UUID and name
+//
+// Deprecated: Use method UUID5.String() instead
+func GenUUID5(ns []byte, name string) string {
+	return UUID5(ns, name).String()
+}
+
+// ////////////////////////////////////////////////////////////////////////////////// //